chore: migrate to commander
