+++ packages/next-codemod/bin/cli.ts
@@ -10,46 +10,17 @@
 
 import globby from 'globby'
 import inquirer from 'inquirer'
-import meow from 'meow'
 import path from 'path'
 import execa from 'execa'
-import { yellow } from 'picocolors'
-import isGitClean from 'is-git-clean'
+// import ciInfo from 'ci-info'
+import { Command } from 'commander'
+// import prompts from 'prompts'
 import { installPackage, uninstallPackage } from '../lib/handle-package'
 import { runUpgrade } from './upgrade'
+import { checkGitStatus, TRANSFORMER_INQUIRER_CHOICES } from '../lib/utils'
 
 export const jscodeshiftExecutable = require.resolve('.bin/jscodeshift')
 export const transformerDirectory = path.join(__dirname, '../', 'transforms')
-
-export function checkGitStatus(force) {
-  let clean = false
-  let errorMessage = 'Unable to determine if git directory is clean'
-  try {
-    clean = isGitClean.sync(process.cwd())
-    errorMessage = 'Git directory is not clean'
-  } catch (err) {
-    if (err && err.stderr && err.stderr.includes('Not a git repository')) {
-      clean = true
-    }
-  }
-
-  if (!clean) {
-    if (force) {
-      console.log(`WARNING: ${errorMessage}. Forcibly continuing.`)
-    } else {
-      console.log('Thank you for using @next/codemod!')
-      console.log(
-        yellow(
-          '\nBut before we continue, please stash or commit your git changes.'
-        )
-      )
-      console.log(
-        '\nYou may use the --force flag to override this safety check.'
-      )
-      process.exit(1)
-    }
-  }
-}
 
 export function runTransform({ files, flags, transformer }) {
   const transformerPath = path.join(transformerDirectory, `${transformer}.js`)
@@ -116,65 +87,6 @@ export function runTransform({ files, flags, transformer }) {
   }
 }
 
-const TRANSFORMER_INQUIRER_CHOICES = [
-  {
-    name: 'name-default-component: Transforms anonymous components into named components to make sure they work with Fast Refresh',
-    value: 'name-default-component',
-  },
-  {
-    name: 'add-missing-react-import: Transforms files that do not import `React` to include the import in order for the new React JSX transform',
-    value: 'add-missing-react-import',
-  },
-  {
-    name: 'withamp-to-config: Transforms the withAmp HOC into Next.js 9 page configuration',
-    value: 'withamp-to-config',
-  },
-  {
-    name: 'url-to-withrouter: Transforms the deprecated automatically injected url property on top level pages to using withRouter',
-    value: 'url-to-withrouter',
-  },
-  {
-    name: 'cra-to-next (experimental): automatically migrates a Create React App project to Next.js',
-    value: 'cra-to-next',
-  },
-  {
-    name: 'new-link: Ensures your <Link> usage is backwards compatible.',
-    value: 'new-link',
-  },
-  {
-    name: 'next-og-import: Transforms imports from `next/server` to `next/og` for usage of Dynamic OG Image Generation.',
-    value: 'next-og-import',
-  },
-  {
-    name: 'metadata-to-viewport-export: Migrates certain viewport related metadata from the `metadata` export to a new `viewport` export.',
-    value: 'metadata-to-viewport-export',
-  },
-  {
-    name: 'next-dynamic-access-named-export: Transforms dynamic imports that return the named export itself to a module like object.',
-    value: 'next-dynamic-access-named-export',
-  },
-  {
-    name: 'next-image-to-legacy-image: safely migrate Next.js 10, 11, 12 applications importing `next/image` to the renamed `next/legacy/image` import in Next.js 13',
-    value: 'next-image-to-legacy-image',
-  },
-  {
-    name: 'next-image-experimental (experimental): dangerously migrates from `next/legacy/image` to the new `next/image` by adding inline styles and removing unused props',
-    value: 'next-image-experimental',
-  },
-  {
-    name: 'built-in-next-font: Uninstall `@next/font` and transform imports to `next/font`',
-    value: 'built-in-next-font',
-  },
-  {
-    name: 'next-async-request-api: Transforms usage of Next.js async Request APIs',
-    value: 'next-async-request-api',
-  },
-  {
-    name: 'next-request-geo-ip: Install `@vercel/functions` to replace `geo` and `ip` properties on `NextRequest`',
-    value: 'next-request-geo-ip',
-  },
-]
-
 function expandFilePathsIfNeeded(filesBeforeExpansion) {
   const shouldExpandFiles = filesBeforeExpansion.some((file) =>
     file.includes('*')
@@ -184,42 +96,49 @@ function expandFilePathsIfNeeded(filesBeforeExpansion) {
     : filesBeforeExpansion
 }
 
-export function run() {
-  const cli = meow({
-    description: 'Codemods for updating Next.js apps.',
-    help: `
-    Usage
-      $ npx @next/codemod <transform> <path> <...options>
-        transform    One of the choices from https://github.com/vercel/next.js/tree/canary/packages/next-codemod
-        path         Files or directory to transform. Can be a glob like pages/**.js
-    Options
-      --force            Bypass Git safety checks and forcibly run codemods
-      --dry              Dry run (no changes are made to files)
-      --print            Print transformed files to your terminal
-      --jscodeshift  (Advanced) Pass options directly to jscodeshift
-    `,
-    flags: {
-      boolean: ['force', 'dry', 'print', 'help'],
-      string: ['_'],
-      alias: {
-        h: 'help',
-      },
-    },
-  } as meow.Options<meow.AnyFlags>)
+const packageJson = require('../package.json')
+const program = new Command(packageJson.name)
+  .description('Codemods for updating Next.js apps.')
+  .version(
+    packageJson.version,
+    '-v, --version',
+    'Output the current version of @next/codemod.'
+  )
+  .argument(
+    '[transform]',
+    'One of the choices from https://github.com/vercel/next.js/tree/canary/packages/next-codemod'
+  )
+  .argument(
+    '[path]',
+    'Files or directory to transform. Can be a glob like pages/**.js'
+  )
+  .usage('[transform] [path] [options]')
+  .helpOption('-h, --help', 'Display this help message.')
+  .option('--force', 'Bypass Git safety checks and forcibly run codemods')
+  .option('--dry', 'Dry run (no changes are made to files)')
+  .option('--print', 'Print transformed files to your terminal')
+  .option('--jscodeshift', '(Advanced) Pass options directly to jscodeshift')
+  .action(run)
+  .allowUnknownOption()
+  .parse(process.argv)
 
-  if (!cli.flags.dry) {
-    checkGitStatus(cli.flags.force)
+const opts = program.opts()
+const { args } = program
+
+function run() {
+  if (!opts.dry) {
+    checkGitStatus(opts.force)
   }
 
-  const isUpgrade = cli.input[0] === 'upgrade' || cli.input[0] === 'up'
+  const isUpgrade = args[0] === 'upgrade' || args[0] === 'up'
 
   if (isUpgrade) {
     return runUpgrade().catch(console.error)
   }
 
   if (
-    cli.input[0] &&
-    !TRANSFORMER_INQUIRER_CHOICES.find((x) => x.value === cli.input[0])
+    args[0] &&
+    !TRANSFORMER_INQUIRER_CHOICES.find((x) => x.value === args[0])
   ) {
     console.error('Invalid transform choice, pick one of:')
     console.error(
@@ -234,7 +153,7 @@ export function run() {
         type: 'input',
         name: 'files',
         message: 'On which files or directory should the codemods be applied?',
-        when: !cli.input[1],
+        when: !args[1],
         default: '.',
         // validate: () =>
         filter: (files) => files.trim(),
@@ -243,7 +162,7 @@ export function run() {
         type: 'list',
         name: 'transformer',
         message: 'Which transform would you like to apply?',
-        when: !cli.input[0],
+        when: !args[0],
         pageSize: TRANSFORMER_INQUIRER_CHOICES.length,
         choices: TRANSFORMER_INQUIRER_CHOICES,
       },
@@ -251,10 +170,10 @@ export function run() {
     .then((answers) => {
       const { files, transformer } = answers
 
-      const filesBeforeExpansion = cli.input[1] || files
+      const filesBeforeExpansion = args[1] || files
       const filesExpanded = expandFilePathsIfNeeded([filesBeforeExpansion])
 
-      const selectedTransformer = cli.input[0] || transformer
+      const selectedTransformer = args[0] || transformer
 
       if (!filesExpanded.length) {
         console.log(`No files found matching ${filesBeforeExpansion.join(' ')}`)
@@ -263,7 +182,7 @@ export function run() {
 
       return runTransform({
         files: filesExpanded,
-        flags: cli.flags,
+        flags: opts,
         transformer: selectedTransformer,
       })
     })
