+++ packages/next-codemod/package.json
@@ -10,13 +10,14 @@
   "dependencies": {
     "chalk": "4.1.2",
     "cheerio": "1.0.0-rc.9",
+    "ci-info": "4.0.0",
+    "commander": "12.1.0",
     "compare-versions": "6.1.1",
     "execa": "4.0.3",
     "globby": "11.0.1",
     "inquirer": "7.3.3",
     "is-git-clean": "1.1.0",
     "jscodeshift": "17.0.0",
-    "meow": "7.0.1",
     "picocolors": "1.0.0",
     "prompts": "2.4.2"
   },
